Fix bug where multiple instances of DarkSplitCorridor34Interpreter would lead to inconstistent State
The internal ref_interpreter was used by all instances (globally). When
one instance called the interpreter and then another, this would lead to
an inconsistent state.

To fix this ref_interpreter is now a member of the interpreter.

Without implementing __rich__ this would lead to huge memory
requirements as the rich method would try to print the whole
ref_interpreter.